machinepdm: support for pluggable PDS/VDS devices.
Processor direct slot (PDS) is the only expansion
slot available in PDM. It accepts a few Apple PDS
cards as well as Nubus cards attached using a special
adapter. The attached device is handled as it would
be a standard Nubus device plugged into the pseudo slot $E.

Video direct slot (VDS) in Cold Fusion is the same
thing as the PDS slot in PDM.
VDS usually carries either the HPV or the Planaria
video expansion card.

Virtual Nubus cards can be attached to the PDS/VDS
slot using the "--pds=[DeviceName]" option.