soundserver_cubeb: fix cross-thread access of the sound out DMA buffer
Instead of directly reading from the sound out DMA buffer in the cubeb
callback, we have a polling task that copies it to a thread-safe ring
buffer (found in the cubeb sources). That buffer can be read from the
cubeb callback safely.

Required removing `using namespace std` in timermanager.h since it was
leading to Windows compilation failures due to ambiguity with the `byte`
type (having `using namespace` in header files is generally discouraged)

Results in compilation failures on windows, and generally discouraged
in header files.